requirements.txt
# Requirements para Análise de Dados - Projeto TechDengue
# Instalar com: pip install -r requirements.txt

# Core - Manipulação de Dados
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pyarrow>=14.0.0
s3fs>=2024.6.0

# Banco de Dados
psycopg2-binary>=2.9.0

# Visualização
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Análise Geoespacial (opcional)
# geopandas>=0.13.0
# folium>=0.14.0
# shapely>=2.0.0

# Machine Learning (opcional)
# scikit-learn>=1.3.0
# statsmodels>=0.14.0
# prophet>=1.1.0

# Dashboard e Web (opcional)
# dash>=2.11.0
# streamlit>=1.24.0

# Jupyter Notebooks (opcional)
# jupyter>=1.0.0
# ipykernel>=6.23.0

# Utilitários
python-dateutil>=2.8.0
pytz>=2023.3
loguru>=0.7.0
pydantic>=2.7,<3
pydantic-settings>=2.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
httpx>=0.28.0

# Cache e Rate Limiting (Fase 1 DaaS)
redis>=5.0.0
slowapi>=0.1.9

# Segurança (Fase 2 DaaS)
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Observabilidade (Fase 5 DaaS)
sentry-sdk[fastapi]>=2.0.0
